+++ eelbrain/plot/__init__.py
@@ -14,8 +14,5 @@ from ._sensors import SensorMaps, SensorMap
 from ._topo import TopoArray, TopoButterfly, Topomap, TopomapBins
 from ._uts import UTSStat, UTS, UTSClusters
 from ._utsnd import Array, Butterfly
-from ._uv import (
-    Barplot, Boxplot, Correlation, Histogram, PairwiseLegend, Regression,
-    Timeplot,
-)
+from ._uv import Barplot, BarplotHorizontal, Boxplot, Correlation, Histogram, PairwiseLegend, Regression, Timeplot
 from . import brain