+++ eelbrain/plot/_uv.py
@@ -3,16 +3,18 @@
 from collections import Sequence
 
 import logging
+from typing import Sequence
 
 import numpy as np
 import scipy.stats  # without this sp.stats is not available
-import matplotlib as mpl
+import matplotlib.axes
 from matplotlib.artist import setp
+import matplotlib as mpl
 
 from .._celltable import Celltable
 from .._data_obj import asvar, ascategorial, assub, cellname
 from .._stats import test, stats
-from ._base import EelFigure, Layout, LegendMixin, YLimMixin, frame_title
+from ._base import EelFigure, Layout, LegendMixin, XAxisMixin, YLimMixin, frame_title
 from ._colors import find_cell_colors
 
 
@@ -234,15 +236,18 @@ class _SimpleFigure(EelFigure):
         self._handles = []
         self._legend = None
 
-    def _set_xlabel_categorial(self, xlabel, model):
-        if xlabel is True:
+    def _set_ax_label_categorial(self, label, model, yaxis=False):
+        if label is True:
             if model and model.name:
-                xlabel = model.name.replace('_', ' ')
+                label = model.name.replace('_', ' ')
             else:
-                xlabel = False
+                label = False
 
-        if xlabel:
-            self._ax.set_xlabel(xlabel)
+        if label:
+            if yaxis:
+                self._ax.set_ylabel(label)
+            else:
+                self._ax.set_xlabel(label)
 
     def _show(self):
         if self._has_frame and not self._layout.w_fixed:
@@ -368,8 +373,8 @@ class Boxplot(YLimMixin, _SimpleFigure):
             raise ValueError(f"colors={colors!r}: need at least as many values as there are cells ({ct.n_cells})")
 
         _SimpleFigure.__init__(self, frame_title(ct.y, ct.x), *args, **kwargs)
-        self._set_xlabel_categorial(xlabel, ct.x)
-        self._configure_yaxis(ct.y, ylabel)
+        self._set_ax_label_categorial(xlabel, ct.x)
+        self._configure_axis(ct.y, ylabel, y=True)
 
         self._plot = p = _plt_boxplot(self._ax, ct, notch, sym, whis, colors, hatch, bottom, top, xticks, xtick_delim, test, par, corr, trend, test_markers, label_fliers)
         p.set_ylim(p.bottom, p.top)
@@ -386,7 +391,7 @@ class Barplot(YLimMixin, _SimpleFigure):
 
     Parameters
     ----------
-    Y : Var
+    y : Var
         Dependent variable.
     x : categorial
         Model (Factor or Interaction).
@@ -471,40 +476,148 @@ class Barplot(YLimMixin, _SimpleFigure):
             pool_error = ct.all_within
 
         _SimpleFigure.__init__(self, frame_title(ct.y, ct.x), *args, **kwargs)
-        self._set_xlabel_categorial(xlabel, ct.x)
-        self._configure_yaxis(ct.y, ylabel)
+        self._set_ax_label_categorial(xlabel, ct.x)
+        self._configure_axis(ct.y, ylabel, y=True)
 
-        p = _plt_barplot(self._ax, ct, error, pool_error, hatch, colors,
-                         bottom, top, origin, None, .5, c, edgec, ec, test, par,
-                         trend, corr, test_markers, xticks, xtick_delim)
+        p = _plt_barplot(self._ax, ct, error, pool_error, hatch, colors, bottom, top, origin, None, .5, c, edgec, ec, test, par, trend, corr, test_markers, xticks, xtick_delim)
         p.set_ylim(p.bottom, p.top)
         p.ax.set_xlim(p.left, p.right)
-        if p.top > 0 > p.bottom:
-            p.ax.axhline(0, color='k')
+        if p.top > p.origin > p.bottom:
+            p.ax.axhline(p.origin, color='k')
 
         YLimMixin.__init__(self, (p,))
         self._show()
 
 
-class _plt_uv_base:
-    """Base for barplot and boxplot"""
+class BarplotHorizontal(XAxisMixin, _SimpleFigure):
+    r"""Barplot for a continuous variable
 
-    def __init__(self, ax, ct, xticks, xtick_delim):
+    Parameters
+    ----------
+    y : Var
+        Dependent variable.
+    x : categorial
+        Model (Factor or Interaction).
+    match : None | categorial
+        Match cases for a repeated measures design.
+    sub : None | index-array
+        Use a subset of the data.
+    cells : None | sequence of cells of x
+        Cells to plot (optional). All entries have to be cells of ``x``). Can be
+        used to change the order of the bars or plot only certain cells.
+    test : bool | scalar
+        True (default): perform pairwise tests;  False: no tests;
+        scalar: 1-sample tests against this value
+    par : bool
+        Use parametric test for pairwise comparisons (use non-parametric
+        tests if False).
+    corr : None | 'hochberg' | 'bonferroni' | 'holm'
+        Method for multiple comparison correction (default 'hochberg').
+    trend : None | str
+        Marker for a trend in pairwise comparisons.
+    test_markers : bool
+        For pairwise tests, plot markers indicating significance level
+        (stars).
+    ylabel : str | None
+        Y axis label (default is inferred from the data).
+    error : str
+        Measure of variability to display in the error bars (default: 1
+        SEM). Examples:
+        'ci': 95% confidence interval;
+        '99%ci': 99% confidence interval (default);
+        '2sem': 2 standard error of the mean.
+    pool_error : bool
+        Pool the errors for the estimate of variability (default is True
+        for related measures designs, False for others). See Loftus & Masson
+        (1994).
+    ec : matplotlib color
+        Error bar color.
+    xlabel : str | bool
+        X axis label (default is ``x.name``).
+    xticks : None | sequence of str
+        X-axis tick labels describing the categories. None to plot no labels
+        (Default uses cell names from ``x``).
+    xtick_delim : str
+        Delimiter for x axis category descriptors (default is ``'\n'``,
+        i.e. the level on each Factor of ``x`` on a separate line).
+    hatch : bool | str
+        Matplotlib Hatch pattern to fill boxes (True to use the module
+        default; default is False).
+    colors : bool | dict | sequence of matplitlib colors
+        Matplotlib colors to use for boxes (True to use the module default;
+        default is False, i.e. no colors).
+    left : scalar
+        Left end of the data axis (default is determined from the data).
+    right : scalar
+        Right end of the data axis (default is determined from the data).
+    origin : scalar
+        Origin of the bars on the data axis (the default is ``0``)
+    c : matplotlib color
+        Bar color (ignored if colors is specified).
+    edgec : matplotlib color
+        Barplot edge color.
+    ds : None | Dataset
+        If a Dataset is specified, all data-objects can be specified as
+        names of Dataset variables
+    frame : bool
+        Draw a frame containing the figure from the top and the right
+        (default ``True``).
+    tight : bool
+        Use matplotlib's tight_layout to resize all axes to fill the figure
+        (default True).
+    ...
+        Also accepts :ref:`general-layout-parameters`.
+    """
+    def __init__(self, y, x=None, match=None, sub=None, cells=None, test=False, par=True,
+                 corr='Hochberg', trend="'", test_markers=True, ylabel=True,
+                 error='sem', pool_error=None, ec='k', xlabel=True, xticks=True,
+                 xtick_delim='\n', hatch=False, colors=False, bottom=0, top=None,
+                 origin=None, c='#0099FF', edgec=None, ds=None, *args, **kwargs):
+        if test is not False:
+            raise NotImplemented("Horizontal barplot with pairwise significance")
+
+        ct = Celltable(y, x, match, sub, cells, ds, asvar)
+
+        if pool_error is None:
+            pool_error = ct.all_within
+
+        _SimpleFigure.__init__(self, frame_title(ct.y, ct.x), *args, **kwargs)
+        self._set_ax_label_categorial(xlabel, ct.x, yaxis=True)
+        self._configure_axis(ct.y, ylabel)
+
+        p = _plt_barplot(self._ax, ct, error, pool_error, hatch, colors,bottom, top, origin, None, .5, c, edgec, ec, test, par, trend, corr, test_markers, xticks, xtick_delim, horizontal=True)
+        p.ax.set_ylim(p.left, p.right)
+        if p.top > p.origin > p.bottom:
+            p.ax.axvline(p.origin, color='k')
+        XAxisMixin.__init__(self, p.bottom, p.top, axes=[p.ax])
+        self._show()
+
+
+class _plt_uv_base:
+    """Base for barplot and boxplot -- x is categorial, y is scalar"""
+
+    def __init__(self, ax, ct, xticks, xtick_delim, horizontal=False):
         # figure decoration
+        axis = ax.yaxis if horizontal else ax.xaxis
         if xticks:
             if xticks is True:
                 xticks = ct.cellnames(xtick_delim)
-            ax.set_xticklabels(xticks)
-            ax.set_xticks(np.arange(len(ct.cells)))
+            axis.set_ticklabels(xticks)
+            axis.set_ticks(np.arange(len(ct.cells)))
         elif xticks is False:
-            ax.set_xticks(())
+            axis.set_ticks(())
 
+        self.horizontal = horizontal
         self.vmin, self.vmax = ax.get_ylim()
         self.ax = ax
 
     def set_ylim(self, bottom, top):
-        self.ax.set_ylim(bottom, top)
-        self.vmin, self.vmax = self.ax.get_ylim()
+        if self.horizontal:
+            self.ax.set_xlim(bottom, top)
+            self.vmin, self.vmax = self.ax.get_xlim()
+        else:
+            self.ax.set_ylim(bottom, top)
+            self.vmin, self.vmax = self.ax.get_ylim()
 
 
 class _plt_boxplot(_plt_uv_base):
@@ -587,24 +700,32 @@ class _plt_boxplot(_plt_uv_base):
 
 
 class _plt_barplot(_plt_uv_base):
-    """Draw a barplot to axes ax for Celltable ct.
-
-    Parameters
-    ----------
-    ax : mpl Axes
-        Axes to which to plot
-    ct : Celltable
-        Data to plot.
-    error : str
-        Variability description (e.g., "95%ci").
-    pool_error : bool
-        Pool the errors for the estimate of variability.
-    ...
-    """
-    def __init__(self, ax, ct, error, pool_error, hatch, colors, bottom, top=None,
-                 origin=None, left=None, width=.5, c='#0099FF', edgec=None,
-                 ec='k', test=True, par=True, trend="'", corr='Hochberg',
-                 test_markers=True, xticks=None, xtick_delim=None):
+    "Barplot from Celltable ``ct``"
+    def __init__(
+            self,
+            ax: mpl.axes.Axes,
+            ct: Celltable,  # Data to plot.
+            error: str,  # Variability description (e.g., "95%ci")
+            pool_error: bool,  # for the variability estimate
+            hatch,
+            colors,
+            bottom,
+            top = None,
+            origin = None,
+            pos: Sequence[float] = None,  # position of the bars
+            width: float = .5,  # width of the pbars
+            c='#0099FF',
+            edgec=None,
+            ec='k',
+            test: bool = True,
+            par=True,
+            trend="'",
+            corr='Hochberg',
+            test_markers=True,
+            xticks=None,
+            xtick_delim=None,
+            horizontal: bool = False,
+    ):
         # kwargs
         if hatch is True:
             hatch = defaults['hatch']
@@ -619,8 +740,8 @@ class _plt_barplot(_plt_uv_base):
 
         # data means
         k = len(ct.cells)
-        if left is None:
-            left = np.arange(k) - width / 2
+        if pos is None:
+            pos = list(range(k))
         height = np.array(ct.get_statistic(np.mean))
 
         # origin
@@ -632,17 +753,19 @@ class _plt_barplot(_plt_uv_base):
             error_match = None
         else:
             error_match = ct.match
-        y_error = stats.variability(ct.y.x, ct.x, error_match, error, pool_error, ct.cells)
+        error_bars = stats.variability(ct.y.x, ct.x, error_match, error, pool_error, ct.cells)
 
         # fig spacing
-        plot_max = np.max(height + y_error)
-        plot_min = np.min(height - y_error)
+        plot_max = np.max(height + error_bars)
+        plot_min = np.min(height - error_bars)
         plot_span = plot_max - plot_min
         y_bottom = min(bottom, plot_min - plot_span * .05)
 
         # main BARPLOT
-        bars = ax.bar(left, height - origin, width, bottom=origin, align='edge',
-                      color=c, edgecolor=edgec, ecolor=ec, yerr=y_error)
+        if horizontal:
+            bars = ax.barh(pos, height - origin, width, origin, color=c, edgecolor=edgec, ecolor=ec, xerr=error_bars)
+        else:
+            bars = ax.bar(pos, height - origin, width, origin, color=c, edgecolor=edgec, ecolor=ec, yerr=error_bars)
 
         # hatch
         if hatch:
@@ -665,11 +788,12 @@ class _plt_barplot(_plt_uv_base):
             ax.axhline(test, color='black')
             y_top = _mark_plot_1sample(ax, ct, par, plot_max, y_unit, test, corr, trend)
 
-        self.left = min(left) - .5 * width
-        self.right = max(left) + 1.5 * width
+        self.left = min(pos) - width
+        self.right = max(pos) + width
+        self.origin = origin
         self.bottom = y_bottom
         self.top = y_top if top is None else top
-        _plt_uv_base.__init__(self, ax, ct, xticks, xtick_delim)
+        _plt_uv_base.__init__(self, ax, ct, xticks, xtick_delim, horizontal)
 
 
 class Timeplot(LegendMixin, YLimMixin, EelFigure):
@@ -791,8 +915,8 @@ class Timeplot(LegendMixin, YLimMixin, EelFigure):
         # get axes
         layout = Layout(1, 1, 5, *args, **kwargs)
         EelFigure.__init__(self, frame_title(y, categories), layout)
-        self._configure_yaxis(y, ylabel)
-        self._configure_xaxis(time, xlabel)
+        self._configure_axis(y, ylabel, y=True)
+        self._configure_axis(time, xlabel)
 
         plot = _ax_timeplot(self._axes[0], y, categories, time, match, colors,
                             hatch, markers, line_plot, error, local_plot,
@@ -862,8 +986,7 @@ class _ax_timeplot:
                         zorder=-999)
                     ax.add_patch(patch)
             elif local_plot == 'bar':
-                _plt_barplot(ax, ct, error, False, hatch, color_list, 0, left=pos,
-                             width=within_spacing, test=False)
+                _plt_barplot(ax, ct, error, False, hatch, color_list, 0, pos=pos, width=within_spacing, test=False)
 
         legend_handles = {}
         if line_plot:
@@ -1003,8 +1126,8 @@ class Correlation(EelFigure, LegendMixin):
         # figure
         layout = Layout(1, 1, 5, *args, autoscale=True, **kwargs)
         EelFigure.__init__(self, frame_title(y, x, cat), layout)
-        self._configure_yaxis(y, ylabel)
-        self._configure_xaxis(x, xlabel)
+        self._configure_axis(y, ylabel, y=True)
+        self._configure_axis(x, xlabel)
 
         ax = self._axes[0]
         legend_handles = {}
@@ -1080,8 +1203,8 @@ class Regression(EelFigure, LegendMixin):
         # figure
         layout = Layout(1, 1, 5, *args, autoscale=True, **kwargs)
         EelFigure.__init__(self, frame_title(y, x, cat), layout)
-        self._configure_xaxis(x, xlabel)
-        self._configure_yaxis(y, ylabel)
+        self._configure_axis(x, xlabel)
+        self._configure_axis(y, ylabel, y=True)
 
         ax = self._axes[0]
         legend_handles = {}
@@ -1258,10 +1381,10 @@ class Histogram(EelFigure):
             xlabel = False  # xlabel is used for test result
 
         if density:
-            self._configure_yaxis('p', 'probability density')
+            self._configure_axis('p', 'probability density', y=True)
         else:
-            self._configure_yaxis('n', 'count')
-        self._configure_xaxis(ct.y, xlabel)
+            self._configure_axis('n', 'count', y=True)
+        self._configure_axis(ct.y, xlabel)
         self._show()